requirements.txt
Flask==1.0.2
discord.py==1.2.3
akinator.py[fast_async]==1.1.3
